# OpenAPI_Swagger.yaml
openapi: 3.0.3
info:
  title: Pet Store API
  description: API para la gestión de productos y servicios de una tienda de mascotas.
  version: 1.0.0
servers:
  - url: https://api.petstore.com/v1
...
paths:
...
  /api/categories:
    get:
      tags:
        - categories
      summary: Listar categorías
      description: Obtiene una lista de todas las categorías de productos.
      operationId: listCategories
      responses:
        '200':
          description: Lista de categorías obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '400':
          description: Error en los parámetros de consulta

    post:
      tags:
        - categories
      summary: Crear una nueva categoría
      description: Permite crear una nueva categoría.
      operationId: createCategory
      requestBody:
        description: Datos de la nueva categoría
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Categoría creada exitosamente
        '400':
          description: Datos de entrada inválidos
  /api/categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        description: ID de la categoría
        required: true
        schema:
          type: string

    get:
      tags:
        - categories
      summary: Obtener detalles de una categoría
      description: Obtiene la información de una categoría específica.
      operationId: getCategoryById
      responses:
        '200':
          description: Detalles de la categoría obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Categoría no encontrada

    put:
      tags:
        - categories
      summary: Actualizar una categoría
      description: Permite a un administrador actualizar los datos de una categoría.
      operationId: updateCategory
      requestBody:
        description: Datos de la categoría a actualizar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Categoría actualizada correctamente
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Categoría no encontrada

    delete:
      tags:
        - categories
      summary: Eliminar una categoría
      description: Permite a un administrador eliminar una categoría.
      operationId: deleteCategory
      responses:
        '200':
          description: Categoría eliminada correctamente
        '404':
          description: Categoría no encontrada
...
  /api/cart:
    get:
      tags:
        - cart
      summary: Obtener carrito actual
      description: Devuelve los productos del carrito de compras actual.
      operationId: getCurrentCart
      responses:
        '200':
          description: Carrito obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          description: Usuario no autenticado

    delete:
      tags:
        - cart
      summary: Vaciar carrito
      description: Permite vaciar todos los productos del carrito actual.
      operationId: emptyCart
      responses:
        '200':
          description: Carrito vaciado exitosamente
        '401':
          description: Usuario no autenticado
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
    CartItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
...